pyproject.toml
[tool.poetry]
name = "globe_tripper_agent"
version = "0.1.0"
description = "multi-model vacation planning Concierge Agent"
authors = ["Izuwa Ahanor izuwa.eng@gmail.com"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.5.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
tenacity = "^9.0.0"
google-adk = "^1.19.0"
google-api-core = "^2.18.0"
google-auth = "^2.29.0"
litellm = "^1.80.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.6"
ruff = "^0.4.2"
mypy = "^1.10.0"
types-PyYAML = "^6.0.12.20240311"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
